Project Setup
The inital project setup for the survival shooter game. This includes a main html file and the js files necessary to run the game. The basic player movement and crosshair aiming has been implemented as well as some of the player's animations. The "assets" folder has been created which contain the sprites, music and sound effects.